remove brackets around species names
It turns out that Artportalen puts brackets around certain taxon
names, even in the Excel export, if the record is marked as 'Ospontan'.
So the label would read [hjärtstilla] or [Primula veris] if the
collector suspected the plant was introduced on the location.

It makes no sense to keep this on the label, so remove such brackets.